@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.17.5 <3.0.0'
+  sdk: '>=2.18.2 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,7 +27,6 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  blue_print_pos: ^0.0.14
   cloud_firestore: ^3.3.0
   cupertino_icons: ^1.0.2
   dio: ^4.0.6
@@ -39,14 +38,16 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_barcode_scanner: ^2.0.0
-  flutter_riverpod: ^1.0.4
-  go_router: ^4.1.0
+  flutter_riverpod: ^2.0.2
+  go_router: ^5.0.5
   google_fonts: ^3.0.1
   image_picker: ^0.8.5+3
   intl: ^0.17.0
   moment_dart: ^0.10.1
   open_file: ^3.2.1
   path_provider: ^2.0.11
+  pdf: ^3.8.3
+  printing: ^5.9.3
   qr_flutter: ^4.0.0
   syncfusion_flutter_xlsio: ^20.2.40-beta
 
